+++ examples/router/index.ts
@@ -5,6 +5,7 @@ import BasicUseage from '../views/BasicUseage.vue'
 import Theme from '../views/Theme.vue'
 import ComponentCustomize from '../views/ComponentCustomize.vue'
 import InScaledBody from '../views/InScaledBody.vue'
+import ChangeContainer from '../views/ChangeContainer.vue'
 
 const routes: Array<RouteRecordRaw> = [
   {
@@ -37,6 +38,11 @@ const routes: Array<RouteRecordRaw> = [
     name: 'InScaledBody',
     component: InScaledBody,
   },
+  {
+    path: '/ChangeContainer',
+    name: 'ChangeContainer',
+    component: ChangeContainer,
+  }
 ]
 
 const router = createRouter({