+++ examples/views/ChangeContainer.vue
@@ -1,23 +1,69 @@
 <template>
   <div>
-    <div class="box1" style="flex:1" @contextmenu="onContextMenu($event)">
-      Right click here to show contextmenu !
+    <div class="horbox">
+      <div class="box1" style="flex: 1;">
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+      </div>
+      <div ref="menuContainer" class="box3" style="flex:4">
+        The menu will be mounted in this container.
+        <div class="box1" style="flex:1;margin:10px 30px" @contextmenu="onContextMenu($event)">
+          Right click to show contextmenu !
+        </div>
+        <div class="box1" style="flex:1;margin:10px 30px" @contextmenu="onContextComponentMenu($event)">
+          And right click to show contextmenu (component mode) !
+        </div>
+      </div>
+      <div class="box1" style="flex: 1;">
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+        long content  long content  long content  long content  long content  long content  long content  long content  long content  long content 
+      </div>
     </div>
+    <div class="box4">
+      You can open examples\views\ChangeContainer.vue file view complete source code.
+    </div>
+    <!--this is component mode of context-menu-->
+    <ContextMenuComponent
+      v-model:show="componentMenuShow"
+      :options="componentMenuOptions"
+    >
+      <context-menu-item label="Simple item" />
+      <context-menu-item label="Item with a icon" icon="icon-reload-1" />
+      <context-menu-group label="Menu with child">
+        <context-menu-item label="Item1" />
+        <context-menu-item label="Item1" />
+      </context-menu-group>
+      <context-menu-sperator />
+      <context-menu-group label="Menu with child child child">
+        <context-menu-item label="Item1" />
+        <context-menu-item label="Item2" />
+        <context-menu-sperator />
+        <context-menu-group label="Child with v-for 50">
+          <context-menu-item v-for="index of 50" :key="index" :label="'Item3-'+index" />
+        </context-menu-group>
+        <context-menu-group label="Childs">
+          <context-menu-item label="Item1-1" />
+          <context-menu-item label="Item1-2" />
+          <context-menu-sperator />
+          <context-menu-group label="Childs">
+            <context-menu-item label="Item2-1" />
+            <context-menu-item label="Item2-2" />
+          </context-menu-group>
+        </context-menu-group>
+      </context-menu-group>
+    </ContextMenuComponent>
   </div>
 </template>
 
 <script setup lang="ts">
-import { onBeforeUnmount, onMounted, reactive } from 'vue'
+import { reactive, ref } from 'vue'
 import { MenuOptions } from '../../src/ContextMenuDefine';
+import ContextMenuComponent from '../../src/ContextMenu.vue';
 import ContextMenu from '../../src/ContextMenuInstance';
 
-onMounted(() => {
-  document.body.setAttribute('style', 'transform: scale(0.5)')
-});
-onBeforeUnmount(() => {
-  document.body.setAttribute('style', '')
-});
-
+const menuContainer = ref<HTMLElement>();
 const menuData = reactive<MenuOptions>({
   items: [
     { 
@@ -165,17 +211,43 @@ const menuData = reactive<MenuOptions>({
   y: 0,
 });
 
+const componentMenuShow = ref(false);
+const componentMenuOptions = reactive<MenuOptions>({
+  getContainer: () => menuContainer.value as HTMLElement,
+  iconFontClass: 'iconfont',
+  customClass: "class-a",
+  zIndex: 3,
+  minWidth: 230,
+  x: 0,
+  y: 0
+});
+
 function onContextMenu(e : MouseEvent) {
   //prevent the browser's default menu
   e.preventDefault();
-  console.log(e);
   
-  const scaledPosition = ContextMenu.transformMenuPosition(e.target as HTMLElement, e.offsetX, e.offsetY);
+  const scaledPosition = ContextMenu.transformMenuPosition(e.target as HTMLElement, e.offsetX, e.offsetY, menuContainer.value);
 
+  menuData.getContainer = menuContainer.value;
   menuData.x = scaledPosition.x;
   menuData.y = scaledPosition.y;
+  console.log(scaledPosition.x, scaledPosition.y);
+
   //show our menu
   ContextMenu.showContextMenu(menuData);
 }
+function onContextComponentMenu(e : MouseEvent) {
+  //prevent the browser's default menu
+  e.preventDefault();
+  
+  const scaledPosition = ContextMenu.transformMenuPosition(e.target as HTMLElement, e.offsetX, e.offsetY, menuContainer.value);
+  componentMenuOptions.x = scaledPosition.x;
+  componentMenuOptions.y = scaledPosition.y;
+  console.log(componentMenuOptions.x, componentMenuOptions.y);
+
+  //show our menu
+  componentMenuShow.value = true;
+}
+
 
 </script>