+++ src/vue-shim.d.ts
@@ -1,7 +1,10 @@
 declare module '*.vue' {
+  import type { DefineComponent } from 'vue'
   import { App, defineComponent } from 'vue'
   const component: ReturnType<typeof defineComponent> & {
     install(app: App): void
   }
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
+  const component: DefineComponent<{}, {}, any>
   export default component
 }